+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.7](https://github.com/searchspring/snap/compare/v0.9.6...v0.9.7) (2021-10-08)
+
+
+### Bug Fixes
+
+* **autocomplete-controller:** added a timeout to ensurethe focus event occurs after click event ([1a23f23](https://github.com/searchspring/snap/commit/1a23f23abca4e4fe8bd7b95973fd1abd4c2ff813))
+* **autocomplete-controller:** removing console logs ([3c73561](https://github.com/searchspring/snap/commit/3c735616314202c44ca9ebeea3a505c48b095518))
+* **snap-preact:** preventing multiple bindings ([d25c4c4](https://github.com/searchspring/snap/commit/d25c4c464cca582aedc7d769a68cf6a741ff9478))
+
+
+
+
+
 ## [0.9.6](https://github.com/searchspring/snap/compare/v0.9.5...v0.9.6) (2021-10-07)
 
 **Note:** Version bump only for package @searchspring/snap